FIX: colours in reports for avg votes
