+++ src/components/charts/teamMoodCircle.tsx
@@ -1,3 +1,4 @@
+import { pickFromGradient } from "@/shared/services/colorService";
 import { HeartIcon, UserGroupIcon } from "@heroicons/react/24/outline";
 import UserIcon from "@heroicons/react/24/outline/UserIcon";
 import { collection, getDocs, getFirestore } from "firebase/firestore";
@@ -26,17 +27,10 @@ const TeamMoodCircle: ComponentType<TeamMoodCircleProps> = ({ teamId }) => {
           }
         });
 
-        const numAvgVotes =
+        const rating =
           votesSnapshot.size > 0 ? totalSentiment / votesSnapshot.size : 0;
-        setNumAvgVotes(numAvgVotes);
-
-        if (numAvgVotes >= 0.8 && numAvgVotes <= 0.9) {
-          setMoodColor(`green`);
-        } else if (numAvgVotes >= 0.4 && numAvgVotes < 0.8) {
-          setMoodColor(`yellow`);
-        } else if (numAvgVotes >= 0.1 && numAvgVotes < 0.4) {
-          setMoodColor(`red`);
-        }
+        setNumAvgVotes(rating);
+        setMoodColor(pickFromGradient(rating));
       };
       getNumVotes();
     }
@@ -45,11 +39,14 @@ const TeamMoodCircle: ComponentType<TeamMoodCircleProps> = ({ teamId }) => {
   return (
     <div>
       <div
-        className={`flex flex-row gap-2 items-center outline outline-2 outline-offset-4 outline-transparent hover:outline-${moodColor}-400 from-${moodColor}-300 to-${moodColor}-500 bg-gradient-to-br text-${moodColor}-900 px-4 py-2 rounded-lg uppercase`}
+        style={{
+          backgroundColor: moodColor,
+        }}
+        className="text-black/50 flex flex-row gap-2 items-center outline outline-2 outline-offset-4 outline-transparent bg-gradient-to-br px-4 py-2 rounded-lg uppercase"
       >
         <UserGroupIcon className="w-6 h-6" />
         <HeartIcon className="w-6 h-6" />
-        {numAvgVotes} {moodColor}
+        {Math.round(numAvgVotes * 100) / 100}
       </div>
     </div>
   );
